Move MvvmManipulatorDemo to ExampleBrowser
