+++ Source/Examples/WPF/ExampleBrowser/ExampleBrowser.csproj
@@ -125,6 +125,10 @@
     <Compile Include="Examples\ModelViewer\Window1.xaml.cs">
       <DependentUpon>Window1.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Examples\MvvmManipulatorDemo\ShellView.xaml.cs">
+      <DependentUpon>ShellView.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Examples\MvvmManipulatorDemo\ShellViewModel.cs" />
     <Compile Include="Examples\PanoramaDemo\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
     </Compile>
@@ -163,6 +167,7 @@
     <Compile Include="Workitems\Workitem10046\MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
     </Compile>
+    <Resource Include="Images\MvvmManipulatorDemo_small.png" />
     <Resource Include="Images\ModelViewer_small.png" />
     <Page Include="Examples\Anaglyph\MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
@@ -245,6 +250,10 @@
       <SubType>Designer</SubType>
     </Page>
     <Page Include="Examples\ModelViewer\Window1.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Examples\MvvmManipulatorDemo\ShellView.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
