+++ Source/Examples/WPF/ExampleBrowser/Examples/MvvmManipulatorDemo/ShellView.xaml.cs
@@ -1,0 +1,20 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="ShellView.xaml.cs" company="Helix Toolkit">
+//   Copyright (c) 2014 Helix Toolkit contributors
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+
+namespace MvvmManipulatorDemo
+{
+    using ExampleBrowser;
+
+    [Example("MVVM manipulation demo", "Translate/Rotate manipulations.")]
+    public partial class ShellView
+    {
+        public ShellView()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
