--- Source/Examples/WPF/MvvmManipulatorDemo/Bootstrapper.cs
@@ -1,12 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="Bootstrapper.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace MvvmManipulatorDemo
-{
-    public class Bootstrapper : Caliburn.Micro.Bootstrapper<MvvmManipulatorDemo.ViewModels.ShellViewModel>
-    {
-    }
-}
