+++ Source/HelixToolkit.Wpf.sln
@@ -25,8 +25,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleBrowser", "Examples\
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Wpf", "Wpf", "{848BE1AD-560A-4419-BD41-994A1A1CCF1F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvvmManipulatorDemo", "Examples\WPF\MvvmManipulatorDemo\MvvmManipulatorDemo.csproj", "{6C9BF028-EDF1-49AC-B441-B886DD807E04}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvvmCameraDemo", "Examples\WPF\MvvmCameraDemo\MvvmCameraDemo.csproj", "{090771E2-3B20-4644-94F3-D128D8D41797}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelixToolkit", "HelixToolkit\HelixToolkit.csproj", "{7C0E987E-CC34-4B19-BA6B-B381AEABB530}"
@@ -124,16 +122,6 @@ Global
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Release|x86.ActiveCfg = Release|Any CPU
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Debug|x86.ActiveCfg = Debug|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Debug|x86.Build.0 = Debug|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Release|Any CPU.ActiveCfg = Release|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Release|Mixed Platforms.Build.0 = Release|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Release|x86.ActiveCfg = Release|x86
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04}.Release|x86.Build.0 = Release|x86
 		{090771E2-3B20-4644-94F3-D128D8D41797}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{090771E2-3B20-4644-94F3-D128D8D41797}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
 		{090771E2-3B20-4644-94F3-D128D8D41797}.Debug|Mixed Platforms.Build.0 = Debug|x86
@@ -175,7 +163,6 @@ Global
 		{47FC8A4C-86B6-4D9D-BE10-EE74F8B992BB} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{991A4C2B-D830-41CB-A810-047ADCF69219} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{848BE1AD-560A-4419-BD41-994A1A1CCF1F} = {A710CED9-33A6-42F1-8A09-55FEC57B6460}
-		{6C9BF028-EDF1-49AC-B441-B886DD807E04} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{090771E2-3B20-4644-94F3-D128D8D41797} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 	EndGlobalSection
 EndGlobal
